# backend/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Client Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenInput"
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
...
components:
  schemas:
...
    RefreshTokenInput:
      type: object
      properties:
        refreshToken:
          type: string
      required: [refreshToken]
...
    AuthResult:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [accessToken, refreshToken, user]

    # ── USER SCHEMAS ─────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: ["string", "null"]
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, developer, tester, client]
        subType:
          type: ["string", "null"]
        designation:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        department:
          type: string
        phoneNumber:
          type: ["string", "null"]
        joiningDate:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, suspended]
        lastLoginAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, email, role, status, createdAt]
...